backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
pydantic
pydantic[email]
passlib[bcrypt]
python-jose